Fix widget instantiation and polymorphism
When instantiating a new widget, only common attributes were initialized
and not the class-specific ones. This has been fixed by using the
registered widgetclass to allocate memory for the widget and then
calling the class-specific allocation function.
This also fixed the polymorphism issue. Thus, it is now possible to cast
an ei_widget_t to a class-specific widget and vice-versa.