Refactor event handling system + Fix bugs
- Separated handle_event into multiple functions for different events
(widget, tag, button click...). This makes the code a lot more readable.
- Fix callback return value handling. If a tag callback returned true,
the event was not propagated to the other widgets even if they matched
the tag. Now, the event is propagated to all widgets matching the tag
and only returns true at the end if at least one callback returned true.
- Fix mouse down event callback function called even if the mouse is not
over the widget. For example, the "pressed" internal callback was binded
to the "button" tag, but if the user clicked on the root widget, the
widgets tree was fully traversed and all buttons were considered as
pressed. Now, a filter_widget was added so that the callback function
would only be called on widgets matching the widget/tag and if it is the
same as the filter_widget (if there is one).
- Add event type check before redrawing the widgets. This is a temporary
fix to avoid redrawing the widgets everytime the user moves the mouse.
This caused issue when moving and cliking on a button too fast (the
click would be registered but the button would not be drawn as pressed
since it is already being drawn by the mouse move event). Once some
optimisations are added, this check will be removed.
- Add ei_app_root_surface implementation