Refactor ei_widget_create() to forbid NULL parent
This commit adds a new function ei_widget_create_internal() which is the
same as ei_widget_create() but without checking if the parent is NULL.
This way, we can create the root widget by calling
ei_widget_create_internal().
The ei_widget_create() now simply checks if the parent is NULL and then
calls ei_widget_create_internal().
This commit also adds a TESTING flag