Update widget creation
Before this commit, the widgets tree was not properly handled. A parent
could only have one child (this had been done for testing purposes). It
is now fixed and and widgets can have children and siblings.
This commit now also uses utility functions to get a size and a rect of
0 instead of hardcoding them.