Refactor placer geom params saving
The code was almost fully duplicated for the case of the widget not
drawn yet and the case already drawn. The only difference was the
allocation/getting of the geom_params. This commit refactors the code to
avoid this duplication and adds functions to get the default values for
each parameter which makes the code a lot more readable.