Update lighten and darken color functions
This commit updates the way the lighten and darken functions work.
Instead of adding/multiplying a set value to the color, we now convert
the color to HSL, modify the lightness and convert it back to RGB, which
works much better for most colors.
See the comment in ei_lighten_color() for more information