Add content_rect and geomnotify implementation
This commit adds the implementation of geomnotify which is called after
a widget's geometry recomputation. It updates the content_rect which,
until now, simply pointed to the screen_location. It is now updated
based on the type of widget to have the correct drawing zone for its
children.
Furthermore, when drawing the children, the clipper is now resized to
take into account the content_rect.