Fix tree traversal
The tree traversal was not working. Instead of iterating over all
widgets, it iterated over all children of a widget and then only the
children of the last child and so on (it was more or less only going
over the right most branch of the tree).
This has been fixed using a recursive preorder traversal algorithm.