Multiple widget's configure function fixes
- Fix an issue where if the widget was already displayed
(ie: handled by a geometry manager, via a call to ei_place for example),
when the user called the configure function of the widget, the size
would be updated in the widget but not in the geom params. This has been
fixed by using the size of the widget instead of the geom param size, if
it's 0, when computing the geometry.
- Fix an issue where a call to configure with NULL parameters
would override the previously set values while it should follow a
"optional parameters" protocol, which it now does.
- Fix an issue where when configuring the text label of a widget, the
label simply pointed to the given text, which could be overriden by
another value at any point, losing the reference to the text. This has
been fixed by allocating memory for the text and copying it using strcpy
when calling configure.