Fix ei_copy_surface wrong resize
This bug was introduced in commit db149b28. When resizing the size of a
size of the rectangles after clipping, it would be set to the biggest
size instead of the smallest one. This was only the case for the width
and not the height. This commit fixes this issue.
It also removes an unused function that had been added in that same
commit