Add custom tags implementation
- Add new tag linked list type
- Add tags linked list field to the widget
- Add functions to add and remove tags from widget. Those tags can be
binded to any function in order to add/remove custom behaviour to/from
widgets
- Fix entry single, double and triple click would all be called in a
row after one click since the event would be handled and then call the
next one. To fix this, new bindings are now added at the beginning of
the list instead of the end. This raised another issue where the triple
click would be before the single click and so, once it would be hanlded,
the single click function would also run. To fix this, a new field has
been added to the entry type
- Fix binding the same event multiple times would add it the list
multiple times
- New bindings are now added at the beginning of the list instead of the
end
- Fix blinking timer would not restart when clicking on the entry
- Fix pressing delete with an active selection in an entry would not
erase the selection but only the character after the selection
- Fix requested size would be reset to the natural size if the configure
function was called with requested size = NULL