Refactor drawing children clipper and copy surface
- Fix children could be drawn outside of their parent. This is because,
when redrawing widgets, the clipper is the invalidated rectangle but
then if a children intersects the clipper, the clipper used to draw this
child is the screen location of the child itself which allowed for any
widget merely intersecting the clippper to be drawn fully (eg: parsing
test application). This has now been fixed by always using the
intersection of the previous clipper and the child's screen location
which means that the clipper can only get smaller
- Fix ei_copy_surface not returning 1 and still copying the surface if
the clipping rectangles are not the same size, which caused issues with
the above mentioned fix. This is now also fixed and the source clipping
rectangle for copying the surface is now clipped based on the
destination clipping rectangle
- Fix toplevel title bar could be drawn used its own size as a clipper
and could thus be drawn outside of the clipper, which erased the text
and close button which were not redrawn because not in the clipper. This
has been fixed by simply intersecting the title bar rect and the clipper