Fix widget natural size not updating when changing
If the user doesn't set the size of the widget explicitally, the widgets
size will default to the natural size of the widget (big enough for its)
properties (border, text, image...). For frames and buttons, if the user
doesn't set the size of the widget but configures an image and later a
rect to crop the image. The widget size won't update to to resize for
the cropped image. This commit fixes this issue, by setting the size of
the widget whenever there is a call to configure except if the size has
been set by the user, in which case, it only resizes the widget if the
size is too small to fit its properties