Add multiline comment support when parsing
- Parsing a file now supports multiline comments, which start with a
'/*'' and end with a '*/'. Multiline comments can be on their own line,
at the end of a line, in the middle of a line (for example, between two
options or two numbers). Multiline comments don't have to be closed if
there are no characters after the comment (ie: the comment is the last
thing before EOF)
- Updated tags_parsing test application to include various multiline
comments